+++ cmd/destroy.go
@@ -36,6 +36,7 @@ func newDestroyCmd() *cobra.Command {
 
 	// Flags for engine.UpdateOptions.
 	var diffDisplay bool
+	var eventLogPath string
 	var parallel int
 	var refresh bool
 	var showConfig bool
@@ -81,6 +82,7 @@ func newDestroyCmd() *cobra.Command {
 				SuppressOutputs:      suppressOutputs,
 				IsInteractive:        interactive,
 				Type:                 displayType,
+				EventLogPath:         eventLogPath,
 				Debug:                debug,
 			}
 
@@ -153,6 +155,9 @@ func newDestroyCmd() *cobra.Command {
 	cmd.PersistentFlags().BoolVar(
 		&diffDisplay, "diff", false,
 		"Display operation as a rich diff showing the overall change")
+	cmd.PersistentFlags().StringVar(
+		&eventLogPath, "event-log", "",
+		"Log events to a file at this path")
 	cmd.PersistentFlags().IntVarP(
 		&parallel, "parallel", "p", defaultParallel,
 		"Allow P resource operations to run in parallel at once (1 for no parallelism). Defaults to unbounded.")
