+++ pkg/backend/display/options.go
@@ -39,5 +39,6 @@ type Options struct {
 	IsInteractive        bool                // true if we should display things interactively.
 	Type                 Type                // type of display (rich diff, progress, or query).
 	JSONDisplay          bool                // true if we should emit the entire diff as JSON.
+	EventLogPath         string              // the path to the file to use for logging events, if any.
 	Debug                bool                // true to enable debug output.
 }
